+++ lib/src/http_request.dart
@@ -3,30 +3,20 @@ import 'dart:typed_data';
 import 'package:http_interop/http_interop.dart';
 
 /// The request which is sent by the client and received by the server
-class HttpRequest extends HttpMessage {
-  HttpRequest(String method, this.uri,
-      {String body = '', Map<String, String> headers = const {}})
-      : method = method.toLowerCase(),
+final class HttpRequest extends HttpMessage {
+  HttpRequest(String method, this.uri, String body,
+      {Map<String, String> headers = const {}})
+      : method = HttpMethod(method),
         super(body, HttpHeaders(headers));
 
-  HttpRequest.binary(String method, this.uri, Uint8List bodyBytes,
+  HttpRequest.binary(String method, this.uri, Uint8List body,
       {Map<String, String> headers = const {}})
-      : method = method.toLowerCase(),
-        super.binary(bodyBytes, HttpHeaders(headers));
+      : method = HttpMethod(method),
+        super.binary(body, HttpHeaders(headers));
 
   /// Requested URI
   final Uri uri;
 
-  /// Request method, lowercase
-  final String method;
-
-  bool get isGet => method == 'get';
-
-  bool get isPost => method == 'post';
-
-  bool get isDelete => method == 'delete';
-
-  bool get isPatch => method == 'patch';
-
-  bool get isOptions => method == 'options';
+  /// Request method.
+  final HttpMethod method;
 }